% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{BinFHEMethod}
\alias{BinFHEMethod}
\title{Binary FHE Methods
Source: binfhe/binfhe-constants.h enum BINFHE_METHOD}
\format{
An object of class \code{list} of length 4.
}
\usage{
BinFHEMethod
}
\description{
Binary FHE Methods
Source: binfhe/binfhe-constants.h enum BINFHE_METHOD
}
\keyword{datasets}
