% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{BinFHEOutput}
\alias{BinFHEOutput}
\title{Binary FHE Output Types
Source: binfhe/binfhe-constants.h enum BINFHE_OUTPUT}
\format{
An object of class \code{list} of length 5.
}
\usage{
BinFHEOutput
}
\description{
Binary FHE Output Types
Source: binfhe/binfhe-constants.h enum BINFHE_OUTPUT
}
\keyword{datasets}
