% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{BinFHEParamSet}
\alias{BinFHEParamSet}
\title{Binary FHE Parameter Sets
Source: binfhe/binfhe-constants.h enum BINFHE_PARAMSET (sequential from 0)}
\format{
An object of class \code{list} of length 44.
}
\usage{
BinFHEParamSet
}
\description{
Binary FHE Parameter Sets
Source: binfhe/binfhe-constants.h enum BINFHE_PARAMSET (sequential from 0)
}
\keyword{datasets}
