% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{BinGate}
\alias{BinGate}
\title{Binary Gate Types
Source: binfhe/binfhe-constants.h enum BINGATE (sequential from 0)}
\format{
An object of class \code{list} of length 14.
}
\usage{
BinGate
}
\description{
Binary Gate Types
Source: binfhe/binfhe-constants.h enum BINGATE (sequential from 0)
}
\keyword{datasets}
