% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{CKKSDataType}
\alias{CKKSDataType}
\title{CKKS Data Type
Source: pke/constants-defs.h enum CKKSDataType}
\format{
An object of class \code{list} of length 2.
}
\usage{
CKKSDataType
}
\description{
CKKS Data Type
Source: pke/constants-defs.h enum CKKSDataType
}
\keyword{datasets}
