% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{DecryptionNoiseMode}
\alias{DecryptionNoiseMode}
\title{Decryption Noise Mode
Source: pke/constants-defs.h enum DecryptionNoiseMode}
\format{
An object of class \code{list} of length 2.
}
\usage{
DecryptionNoiseMode
}
\description{
Decryption Noise Mode
Source: pke/constants-defs.h enum DecryptionNoiseMode
}
\keyword{datasets}
