% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{DistributionType}
\alias{DistributionType}
\title{Distribution Type (lattice parameters)
Source: core/lattice/stdlatticeparms.h enum DistributionType}
\format{
An object of class \code{list} of length 3.
}
\usage{
DistributionType
}
\description{
Distribution Type (lattice parameters)
Source: core/lattice/stdlatticeparms.h enum DistributionType
}
\keyword{datasets}
