% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{EncryptionTechnique}
\alias{EncryptionTechnique}
\title{Encryption Technique
Source: pke/constants-defs.h enum EncryptionTechnique}
\format{
An object of class \code{list} of length 2.
}
\usage{
EncryptionTechnique
}
\description{
Encryption Technique
Source: pke/constants-defs.h enum EncryptionTechnique
}
\keyword{datasets}
