% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{ExecutionMode}
\alias{ExecutionMode}
\title{Execution Mode
Source: pke/constants-defs.h enum ExecutionMode}
\format{
An object of class \code{list} of length 2.
}
\usage{
ExecutionMode
}
\description{
Execution Mode
Source: pke/constants-defs.h enum ExecutionMode
}
\keyword{datasets}
