% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{Feature}
\alias{Feature}
\title{PKE Scheme Features (bitmask)
Source: pke/constants-defs.h enum PKESchemeFeature}
\format{
An object of class \code{list} of length 8.
}
\usage{
Feature
}
\description{
PKE Scheme Features (bitmask)
Source: pke/constants-defs.h enum PKESchemeFeature
}
\keyword{datasets}
