% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{KeygenMode}
\alias{KeygenMode}
\title{Key Generation Mode
Source: binfhe/binfhe-constants.h enum KEYGEN_MODE}
\format{
An object of class \code{list} of length 2.
}
\usage{
KeygenMode
}
\description{
Key Generation Mode
Source: binfhe/binfhe-constants.h enum KEYGEN_MODE
}
\keyword{datasets}
