% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{KeySwitchTechnique}
\alias{KeySwitchTechnique}
\title{Key Switching Techniques
Source: pke/constants-defs.h enum KeySwitchTechnique}
\format{
An object of class \code{list} of length 3.
}
\usage{
KeySwitchTechnique
}
\description{
Key Switching Techniques
Source: pke/constants-defs.h enum KeySwitchTechnique
}
\keyword{datasets}
