% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{MultiplicationTechnique}
\alias{MultiplicationTechnique}
\title{Multiplication Technique (BFV)
Source: pke/constants-defs.h enum MultiplicationTechnique}
\format{
An object of class \code{list} of length 4.
}
\usage{
MultiplicationTechnique
}
\description{
Multiplication Technique (BFV)
Source: pke/constants-defs.h enum MultiplicationTechnique
}
\keyword{datasets}
