% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openfhe-package.R
\docType{package}
\name{openfhe-package}
\alias{openfhe}
\alias{openfhe-package}
\title{openfhe: Fully Homomorphic Encryption via the 'OpenFHE' Library}
\description{
Compute directly on encrypted data using fully homomorphic encryption, without ever decrypting it. Provides high level idiomatic interfaces to the 'OpenFHE' library (Al Badawi et al., 2022) \doi{10.1145/3560810.3564267}, which is bundled, including the 'BFV' (Fan and Vercauteren, 2012) \url{https://eprint.iacr.org/2012/144}, 'BGV' (Brakerski, Gentry and Vaikuntanathan, 2014) \doi{10.1145/2633600} and 'CKKS' (Cheon et al., 2017) \doi{10.1007/978-3-319-70694-8_15} schemes, with support for encrypted vectors and matrices, arithmetic and linear algebra operators, rotations and reductions, approximate evaluation of smooth functions, encrypted comparisons via scheme switching, 'CKKS' bootstrapping, and serialisation of keys and ciphertexts.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/louisaslett/openfhe}
  \item Report bugs at \url{https://github.com/louisaslett/openfhe/issues}
}

}
\author{
\strong{Maintainer}: Louis Aslett \email{louis.aslett@durham.ac.uk} (\href{https://orcid.org/0000-0003-2211-233X}{ORCID}) [copyright holder]

Authors:
\itemize{
  \item Louis Aslett \email{louis.aslett@durham.ac.uk} (\href{https://orcid.org/0000-0003-2211-233X}{ORCID}) [copyright holder]
}

Other contributors:
\itemize{
  \item Duality Technologies Inc. (Bundled OpenFHE library) [copyright holder]
  \item New Jersey Institute of Technology (NJIT) (Bundled OpenFHE library) [copyright holder]
  \item The OpenFHE contributors (Bundled OpenFHE library; see inst/COPYRIGHTS) [copyright holder]
  \item Samuel Neves (Bundled BLAKE2 implementation) [copyright holder]
  \item W. Shane Grant (Bundled 'cereal' serialisation library) [copyright holder]
  \item Randolph Voorhies (Bundled 'cereal' serialisation library) [copyright holder]
}

}
\keyword{internal}
