% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{PlaintextEncodings}
\alias{PlaintextEncodings}
\title{Plaintext Encoding Types
Source: pke/constants-defs.h enum PlaintextEncodings}
\format{
An object of class \code{list} of length 5.
}
\usage{
PlaintextEncodings
}
\description{
Plaintext Encoding Types
Source: pke/constants-defs.h enum PlaintextEncodings
}
\keyword{datasets}
