% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{ScalingTechnique}
\alias{ScalingTechnique}
\title{Scaling Techniques (CKKS)
Source: pke/constants-defs.h enum ScalingTechnique}
\format{
An object of class \code{list} of length 8.
}
\usage{
ScalingTechnique
}
\description{
Scaling Techniques (CKKS)
Source: pke/constants-defs.h enum ScalingTechnique
}
\keyword{datasets}
