% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{SecretKeyDist}
\alias{SecretKeyDist}
\title{Secret Key Distribution
Source: core/lattice/constants-lattice.h enum SecretKeyDist}
\format{
An object of class \code{list} of length 4.
}
\usage{
SecretKeyDist
}
\description{
Secret Key Distribution
Source: core/lattice/constants-lattice.h enum SecretKeyDist
}
\keyword{datasets}
