% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enums.R
\docType{data}
\name{SecurityLevel}
\alias{SecurityLevel}
\title{Security Levels
Source: core/lattice/stdlatticeparms.h enum SecurityLevel}
\format{
An object of class \code{list} of length 7.
}
\usage{
SecurityLevel
}
\description{
Security Levels
Source: core/lattice/stdlatticeparms.h enum SecurityLevel
}
\keyword{datasets}
