% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\docType{class}
\name{Ciphertext-class}
\alias{Ciphertext-class}
\alias{show,Ciphertext-method}
\alias{length,Ciphertext-method}
\title{Encrypted vector}
\usage{
\S4method{show}{Ciphertext}(object)

\S4method{length}{Ciphertext}(x)
}
\arguments{
\item{object}{A \code{Ciphertext}.}

\item{x}{A \code{Ciphertext}.}
}
\description{
A \code{Ciphertext} is one encrypted R vector: all of its elements are packed
into the SIMD slots of a single OpenFHE ciphertext.  Create one with
\code{\link[=encrypt]{encrypt()}}; recover the plaintext with \code{\link[=decrypt]{decrypt()}}.  The true R length
is recorded so that decryption never exposes slot padding.
}
\section{Functions}{
\itemize{
\item \code{show(Ciphertext)}: Compact description of an encrypted vector:
its length, scheme and multiplicative level.

\item \code{length(Ciphertext)}: The true length of the encrypted vector.

}}
\section{Slots}{

\describe{
\item{\code{ptr}}{External pointer to the native OpenFHE ciphertext.}

\item{\code{context}}{The \link[=FHEContext-class]{FHEContext} the ciphertext was encrypted under.}

\item{\code{length}}{The true length of the encrypted vector.}
}}

\seealso{
\code{\link[=encrypt]{encrypt()}}, \code{\link[=decrypt]{decrypt()}}
}
