% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openfhe-package.R
\docType{package}
\name{openfhe-package}
\alias{openfhe}
\alias{openfhe-package}
\title{openfhe: R Interface to the OpenFHE Fully Homomorphic Encryption Library}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides an R interface to the OpenFHE C++ library for fully homomorphic encryption (FHE). Supports BFV, BGV, and CKKS schemes for computation on encrypted data, as well as FHEW/TFHE for Boolean circuit evaluation. Designed as a production-grade upgrade path from the 'homomorpheR' package.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://bnaras.github.io/openfhe}
  \item \url{https://github.com/bnaras/fhe}
  \item Report bugs at \url{https://github.com/bnaras/fhe/issues}
}

}
\author{
\strong{Maintainer}: Balasubramanian Narasimhan \email{naras@stanford.edu} (\href{https://orcid.org/0000-0001-5852-7639}{ORCID})

}
